% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaching_table.R
\name{coach_main_table}
\alias{coach_main_table}
\title{Coaching Table}
\usage{
coach_main_table(coach, fwd)
}
\arguments{
\item{coach}{coach}

\item{fwd}{database}
}
\value{
datatable
}
\description{
generates coaching table
}
\examples{
coach_main_table(coach,fwd)

}
