% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaching_table.R
\name{coaching_table}
\alias{coaching_table}
\title{Coaching Table}
\usage{
coaching_table(coach, s)
}
\arguments{
\item{s}{}
}
\description{
generates coaching table
}
\examples{
coaching_table(coach,s)

}
