% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_title.R
\name{create_title}
\alias{create_title}
\title{Create Title}
\usage{
create_title(pA = NULL, pB = NULL, pC = NULL)
}
\arguments{
\item{pC}{}
}
\description{
creates plot title
}
\examples{
create_title(pA=NULL,pB=NULL,pC=NULL)

}
