% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaching_plot.R
\name{get_coach_positional_chart}
\alias{get_coach_positional_chart}
\title{Coaching Plot}
\usage{
get_coach_positional_chart(coach, position, stat_name, fwd, output_type)
}
\arguments{
\item{coach}{coach}

\item{position}{position}

\item{stat_name}{stat name}

\item{fwd}{database}

\item{output_type}{table or chart}
}
\value{
plotly object or datatable
}
\description{
creates coach table or plot
}
\examples{
get_coach_positional_chart(coach,position,stat_name,fwd,output_type)

}
