% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dfs_platform_salaries_for_opt.R
\name{get_dfs_platform_salaries}
\alias{get_dfs_platform_salaries}
\title{DFS Platform Salary Data}
\usage{
get_dfs_platform_salaries(con, dfs_platform, platforms_list)
}
\arguments{
\item{con}{RPostgres connection object pointed at the database.}

\item{dfs_platform}{DFS platform string: DraftKings, FandDuel, Yahoo!}

\item{platforms_list}{list of DFS platform attributes: keys}
}
\description{
function to pull DFS Platform Salary data
}
\examples{
get_dfs_platform_salaries_for_opt('DrafKings',
 list(DraftKings=list(proj_points='dk_proj_points', salary='dksalary', id='dkid')))

}
