% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_player_table.R
\name{get_player_stats}
\alias{get_player_stats}
\title{College Player Stats}
\usage{
get_player_stats(player, con, datatype)
}
\arguments{
\item{player}{player names}

\item{con}{DB connection}

\item{datatype}{gamelogs or year}
}
\value{
dataframe
}
\description{
generates college stats table
}
\examples{
get_player_stats(player, con, datatype)

}
