% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_table.R
\name{get_qb_positional_chart}
\alias{get_qb_positional_chart}
\title{QB Plot}
\usage{
get_qb_positional_chart(qb, position, stat_name, fwd, pfrplayers)
}
\arguments{
\item{qb}{qb}

\item{position}{position}

\item{stat_name}{stat name}

\item{fwd}{database}

\item{pfrplayers}{database}
}
\value{
plotly object
}
\description{
creates qb plot
}
\examples{
get_qb_positional_chart(qb,position,stat_name,fwd)

}
