% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_splits_chart.R
\name{home_splits_chart}
\alias{home_splits_chart}
\title{Home Splits Chart}
\usage{
home_splits_chart(players, years, scoring, fwd, pfrplayers)
}
\arguments{
\item{players}{player name}

\item{years}{year vector}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{fwd}{a player database}

\item{pfrplayers}{PFR database}
}
\value{
plotly object
}
\description{
creates home splits chart
}
\examples{
home_splits_chart(players,years,scoring,fwd,pfrplayers)

}
