% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_splits_table.R
\name{home_splits_table}
\alias{home_splits_table}
\title{Home Splits Table}
\usage{
home_splits_table(players, years, scoring, fwd, pfrplayers)
}
\arguments{
\item{players}{player name}

\item{years}{year vector}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{fwd}{a player database}

\item{pfrplayers}{PFR database}
}
\value{
datatable
}
\description{
creates home splits table
}
\examples{
home_splits_table(players,years,scoring,fwd,pfrplayers)

}
