% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiszero.R
\name{na.is.zero}
\alias{na.is.zero}
\title{NA.Is.Zero}
\usage{
na.is.zero(X, value = 0)
}
\arguments{
\item{value}{set to 0}
}
\value{

}
\description{
na to zero
}
\examples{
na.is.zero(X,value=)

}
