% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_lineup.R
\name{optimize_lineup}
\alias{optimize_lineup}
\title{Lineup Optimizer}
\usage{
optimize_lineup(dat, budget_in, qb_fixed, rb_fixed, wr_fixed, te_fixed,
  dst_fixed)
}
\arguments{
\item{dat}{DFS datatable}

\item{budget_in}{budget price}

\item{qb_fixed}{QB to optimize around}

\item{rb_fixed}{RB to optimize around}

\item{wr_fixed}{WR to optimize around}

\item{te_fixed}{TE to optimize around}

\item{dst_fixed}{DST to optimize around}
}
\value{
datatable
}
\description{
optimizes DFS lineup
}
\examples{
optimize_lineup(dat, budget_in, qb_fixed, rb_fixed, wr_fixed, te_fixed, dst_fixed)

}
