% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_lineup_get_input_file.R
\name{optimize_lineup_get_input_file}
\alias{optimize_lineup_get_input_file}
\title{Lineup Optimizer Get Input FIle}
\usage{
optimize_lineup_get_input_file(input)
}
\arguments{
\item{input}{}
}
\value{
datatable
}
\description{
gets input file for optimizer
}
\examples{
optimize_lineup_get_input_file(input)

}
