% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_table.R
\name{points_table}
\alias{points_table}
\title{Points Table}
\usage{
points_table(position, years, con, pfrplayers, minpassatt, minrushattQB,
  minrushattRB, mintargetRB, mintargetWR, mintargetTE)
}
\arguments{
\item{position}{position}

\item{years}{a vector of years}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}

\item{minpassatt}{number}

\item{minrushattQB}{number}

\item{minrushattRB}{number}

\item{mintargetRB}{number}

\item{mintargetWR}{number}

\item{mintargetTE}{number}
}
\value{
dataframe
}
\description{
generates points table
}
\examples{
points_table(position,years,con,pfrplayers,minpassatt,minrushattQB,minrushattRB,mintargetRB,mintargetWR,mintargetTE)

}
