% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posrank_chart.R
\name{posrank_chart}
\alias{posrank_chart}
\title{Position Rank Chart}
\usage{
posrank_chart(players, years, scoring, con, pfrplayers)
}
\arguments{
\item{players}{a vector of players}

\item{years}{a vector of years}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
plotly object
}
\description{
creates posrank chart
}
\examples{
posrank_chart(players,years,scoring,con,pfrplayers)

}
