% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppg_chart.R
\name{ppg_chart}
\alias{ppg_chart}
\title{PPG Chart}
\usage{
ppg_chart(players, years, scoring, con, pfrplayers)
}
\arguments{
\item{players}{a vector of players}

\item{years}{a vector of years}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
plotly object
}
\description{
creates ppg chart
}
\examples{
ppg_chart(players,years,scoring,con,pfrplayers)

}
