% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppg_table.R
\name{ppg_table}
\alias{ppg_table}
\title{PPG Table}
\usage{
ppg_table(players, years, scoring, con, pfrplayers)
}
\arguments{
\item{players}{a vector of players}

\item{years}{a vector of years}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
dataframe
}
\description{
creates ppg table
}
\examples{
ppg_table(players,years,scoring,con,pfrplayers)

}
