% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_raw_table.R
\name{ppo_raw_table}
\alias{ppo_raw_table}
\title{PPO Raw Table}
\usage{
ppo_raw_table(years, scoring, con, pfrplayers)
}
\arguments{
\item{years}{a vector of years}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
datatable
}
\description{
creates ppo raw table
}
\examples{
ppo_raw_table(years,scoring,con,pfrplayers)

}
