% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_receiving_chart.R
\name{ppo_receiving_chart}
\alias{ppo_receiving_chart}
\title{PPO Receiving Plot}
\usage{
ppo_receiving_chart(players, years, scoring, con, pfrplayers)
}
\arguments{
\item{players}{a vector of players}

\item{years}{a vector of years}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
plotly object
}
\description{
creates receiving ppo plot
}
\examples{
ppo_receiving_chart(players,years,scoring,con,pfrplayers)

}
