% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppo_table.R
\name{ppo_table}
\alias{ppo_table}
\title{PPO Table}
\usage{
ppo_table(players, years, scoring, con, pfrplayers)
}
\arguments{
\item{players}{a vector of players}

\item{years}{a vector of years}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
datatable
}
\description{
creates ppo table
}
\examples{
ppo_table(players,years,scoring,con,pfrplayers)

}
