% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_splits.R
\name{team_splits}
\alias{team_splits}
\title{Team Splits Table}
\usage{
team_splits(player, years, pfrplayers)
}
\arguments{
\item{player}{a player name}

\item{years}{year vector}

\item{pfrplayers}{PFR database}
}
\value{
datatable
}
\description{
creates team splits table
}
\examples{
team_splits(player,years,pfrplayers)

}
