% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_range.R
\name{week_range}
\alias{week_range}
\title{Stats by Week Range Table}
\usage{
week_range(stat, years, weeks, con)
}
\arguments{
\item{stat}{a football stat}

\item{years}{a vector of years}

\item{weeks}{vector of weeks}

\item{con}{DB connection}
}
\value{
datatable
}
\description{
creates stats by week range table
}
\examples{
week_range(stat,years,weeks,con)

}
