% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_splits.R
\name{weekly_split}
\alias{weekly_split}
\title{Weekly Splits Table}
\usage{
weekly_split(fwd, player_pos_name, measure_vars, year_filter, week_split,
  player_pos_split, home_away_split, team_split, opponent_split,
  gameday_split, gamestart_time_split, win_split, posrank_year = 2018,
  df_output, transpose_output = TRUE, pfrplayers)
}
\arguments{
\item{fwd}{player database}

\item{player_pos_name}{player to check splits}

\item{measure_vars}{variables to check splits}

\item{year_filter}{vector of years}

\item{week_split}{vector of weeks}

\item{player_pos_split}{player to split on}

\item{team_split}{team to split on}

\item{opponent_split}{opponent to split on}

\item{gameday_split}{game day to split on}

\item{gamestart_time_split}{game time to split on}

\item{win_split}{win/loss to split on}

\item{posrank_year}{year to get posrank for projection}

\item{df_output}{dataframe to return}

\item{transpose_output}{default to TRUE}

\item{pfrplayers}{PFR player database}
}
\value{
dataframe
}
\description{
creates splits table
}
\examples{
weekly_split(fwd,player_pos_name, measure_vars, year_filter, week_split, player_pos_split, home_away_split, team_split,opponent_split, gameday_split, gamestart_time_split, win_split,posrank_year = 2018, df_output, transpose_output = TRUE,pfrplayers)

}
