% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_table.R
\name{yearly_comp_player_df}
\alias{yearly_comp_player_df}
\title{Yearly Table}
\usage{
yearly_comp_player_df(pA = NULL, pB = NULL, pC = NULL, years,
  scoring)
}
\arguments{
\item{scoring}{}
}
\description{
generates yearly table
}
\examples{
yearly_comp_player_df(pA=NULL,pB=NULL,pC=NULL,years,scoring)

}
