% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_plot.R
\name{yearly_comp_player_ggplot2}
\alias{yearly_comp_player_ggplot2}
\title{Yearly Plot ggplot}
\usage{
yearly_comp_player_ggplot2(pA = NULL, pB = NULL, pC = NULL, years,
  scoring)
}
\arguments{
\item{scoring}{}
}
\description{
creates yearly plot
}
\examples{
yearly_comp_player_ggplot2(pA=NULL,pB=NULL,pC=NULL,years,scoring)

}
