% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_plot.R
\name{yearly_player_chart}
\alias{yearly_player_chart}
\title{Yearly Plot}
\usage{
yearly_player_chart(players, years, scoring, con)
}
\arguments{
\item{con}{}
}
\description{
creates yearly plot
}
\examples{
yearly_player_chart <- function(players,years,scoring,con)

}
