% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_table.R
\name{yearly_player_df}
\alias{yearly_player_df}
\title{Yearly Table}
\usage{
yearly_player_df(players, years, con, pfrplayers)
}
\arguments{
\item{players}{a vector of players}

\item{years}{a vector of years}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
dataframe
}
\description{
generates yearly table
}
\examples{
yearly_player_df(players,years,con,pfrplayers)

}
