% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_chart.R
\name{zscore_chart}
\alias{zscore_chart}
\title{Z-Score Chart}
\usage{
zscore_chart(players, position, scoring)
}
\arguments{
\item{players}{player name}

\item{position}{QB,RB,WR,TE}

\item{scoring}{ppr,halfppr,standard,sixpttd}
}
\value{
plotly object
}
\description{
creates z-score chart
}
\examples{
zscore_chart(players,position,scoring)

}
