% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_plot.R
\name{yearly_player_chart}
\alias{yearly_player_chart}
\title{Yearly Plot}
\usage{
yearly_player_chart(players, years, scoring, con, pfrplayers)
}
\arguments{
\item{players}{a vector of players}

\item{years}{a vector of years}

\item{scoring}{ppr,halfppr,standard,sixpttd}

\item{con}{DB connection}

\item{pfrplayers}{pfr DB}
}
\value{
plotly object
}
\description{
creates yearly plot
}
\examples{
yearly_player_chart(players,years,scoring,con,pfrplayers)

}
