% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_table.R
\name{yearly_player_df}
\alias{yearly_player_df}
\title{Yearly Table}
\usage{
yearly_player_df(players, years, con)
}
\arguments{
\item{con}{}
}
\description{
generates yearly table
}
\examples{
yearly_comp_player_df(players,years,con)

}
